Improve usage borg list
1. Use borg 1.1 list features for easier processing. With --format and
of --last 1 of borg list, it's now possible to let determine borg,
which archive is the current one and it spits out the time directly.

2. Borg 1.0 never supported list --format for archives. It may be the
case, that I developed this plugin in combination with a self-installed
pip beta executable. There is now a fallback introduced, which gets the
time from the latest archive also under borg 1.0. It's ugly but works
and only meant as a fallback.